Extract settings from TPlayer
