+++ src/main/java/me/makkuusen/timing/system/gui/ColorSettingsGui.java
@@ -2,7 +2,7 @@ package me.makkuusen.timing.system.gui;
 
 import me.makkuusen.timing.system.ApiUtilities;
 import me.makkuusen.timing.system.ItemBuilder;
-import me.makkuusen.timing.system.TPlayer;
+import me.makkuusen.timing.system.tplayer.TPlayer;
 import me.makkuusen.timing.system.sounds.PlaySound;
 import me.makkuusen.timing.system.theme.Text;
 import me.makkuusen.timing.system.theme.messages.Gui;
@@ -31,7 +31,7 @@ public class ColorSettingsGui extends BaseGui {
         var hexColor = ApiUtilities.getHexFromDyeColor(dye);
         var button = new GuiButton(new ItemBuilder(dye).setName(Text.get(tPlayer, Gui.COLOR).color(TextColor.fromHexString(hexColor))).build());
         button.setAction(() -> {
-            tPlayer.setHexColor(hexColor);
+            tPlayer.getSettings().setHexColor(hexColor);
             PlaySound.buttonClick(tPlayer);
             new SettingsGui(tPlayer).show(tPlayer.getPlayer());
         });
