+++ src/main/java/me/makkuusen/timing/system/heat/DriverScoreboard.java
@@ -1,6 +1,6 @@
 package me.makkuusen.timing.system.heat;
 
-import me.makkuusen.timing.system.TPlayer;
+import me.makkuusen.timing.system.tplayer.TPlayer;
 import me.makkuusen.timing.system.TimingSystem;
 import me.makkuusen.timing.system.participant.Driver;
 import me.makkuusen.timing.system.round.QualificationRound;
@@ -28,7 +28,7 @@ public class DriverScoreboard {
 
     public void setTitle() {
         String eventName;
-        if (tPlayer.getCompactScoreboard() && heat.getEvent().getDisplayName().length() > 8) {
+        if (tPlayer.getSettings().getCompactScoreboard() && heat.getEvent().getDisplayName().length() > 8) {
             eventName = heat.getEvent().getDisplayName().substring(0, 8);
         } else {
             eventName = heat.getEvent().getDisplayName();
@@ -71,10 +71,10 @@ public class DriverScoreboard {
                 break;
             }
             if (heat.getRound() instanceof QualificationRound) {
-                lines.add(getDriverRowQualification(driver, this.driver, tPlayer.getCompactScoreboard(), tPlayer.getTheme()));
+                lines.add(getDriverRowQualification(driver, this.driver, tPlayer.getSettings().getCompactScoreboard(), tPlayer.getTheme()));
 
             } else {
-                lines.add(getDriverRowFinal(driver, this.driver, tPlayer.getCompactScoreboard(), tPlayer.getTheme()));
+                lines.add(getDriverRowFinal(driver, this.driver, tPlayer.getSettings().getCompactScoreboard(), tPlayer.getTheme()));
             }
         }
         return lines;
@@ -90,10 +90,10 @@ public class DriverScoreboard {
                 break;
             }
             if (heat.getRound() instanceof QualificationRound) {
-                lines.add(getDriverRowQualification(driver, this.driver, tPlayer.getCompactScoreboard(), tPlayer.getTheme()));
+                lines.add(getDriverRowQualification(driver, this.driver, tPlayer.getSettings().getCompactScoreboard(), tPlayer.getTheme()));
 
             } else {
-                lines.add(getDriverRowFinal(driver, this.driver, tPlayer.getCompactScoreboard(), tPlayer.getTheme()));
+                lines.add(getDriverRowFinal(driver, this.driver, tPlayer.getSettings().getCompactScoreboard(), tPlayer.getTheme()));
             }
         }
         return lines;
