+++ src/main/java/me/makkuusen/timing/system/participant/Subscriber.java
@@ -1,7 +1,7 @@
 package me.makkuusen.timing.system.participant;
 
 import co.aikar.idb.DbRow;
-import me.makkuusen.timing.system.TPlayer;
+import me.makkuusen.timing.system.tplayer.TPlayer;
 import me.makkuusen.timing.system.database.TSDatabase;
 
 import java.util.UUID;
