+++ src/main/java/me/makkuusen/timing/system/theme/Text.java
@@ -1,7 +1,7 @@
 package me.makkuusen.timing.system.theme;
 
 import com.destroystokyo.paper.ClientOption;
-import me.makkuusen.timing.system.TPlayer;
+import me.makkuusen.timing.system.tplayer.TPlayer;
 import me.makkuusen.timing.system.TimingSystem;
 import me.makkuusen.timing.system.theme.messages.ActionBar;
 import me.makkuusen.timing.system.theme.messages.Message;
