+++ src/main/java/me/makkuusen/timing/system/timetrial/TimeTrialAttempt.java
@@ -2,7 +2,7 @@ package me.makkuusen.timing.system.timetrial;
 
 import co.aikar.idb.DbRow;
 import lombok.Getter;
-import me.makkuusen.timing.system.TPlayer;
+import me.makkuusen.timing.system.tplayer.TPlayer;
 import me.makkuusen.timing.system.database.TSDatabase;
 
 import java.util.UUID;
