+++ src/main/java/me/makkuusen/timing/system/track/TimeTrials.java
@@ -3,7 +3,7 @@ package me.makkuusen.timing.system.track;
 import co.aikar.idb.DB;
 import lombok.Getter;
 import me.makkuusen.timing.system.ApiUtilities;
-import me.makkuusen.timing.system.TPlayer;
+import me.makkuusen.timing.system.tplayer.TPlayer;
 import me.makkuusen.timing.system.TimingSystem;
 import me.makkuusen.timing.system.timetrial.TimeTrialAttempt;
 import me.makkuusen.timing.system.timetrial.TimeTrialFinish;
